Use release-plz as the release tool (#43)
https://release-plz.dev/

It adds one GH action with two jobs that are triggered for every commit
to `main`
* `release-plz-release`: it will release the package version if it was
not released to crates.io yet (it uses @jgsogo token secret).
**Note.-**, when this PR is merged (if everything is right) the first
version of the crate will be released (closes #39).
* `release-plz-pr`: keeps a PR updated with all the changes that are
being accumulated for the next release. If using [conventional
commits](https://www.conventionalcommits.org/), it will also update the
next semver version in this PR according to the changes.

---

I've also added here the `dependabot.yml` to update the GH actions
version automatically.

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>