perf: Order `TargetingRule`s in properties and features just once (#50)
(Warming up after Christmas)

Targeting/Segment rules are no longer ordered every time they are
evaluated. Now the Configuration object performs the ordering just once
when it is created.

My guess is that a property/feature is evaluated more times than it is
created, so the cost should be amortized.

---------

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>