[refact] Introduce NetworkResult and NetworkError types (#64)
Extracted from #60
Working on #63

In this PR we are just introducing new types `NetworkResult` and
`NetworkError` that will be used inside the `network` module. Our
purpose is to isolate inside this `network` module everything related to
outside-world communications and provide a clear (and small) interface
that can be reused from the other modules in the crate.

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>