Move `Configuration` sync to `network::live_configuration` module, outside `AppConfigurationHttp` object (#66)
Taken from #60 -- This PR basically extracts everything related to the
refactor, and **leaves the offline modes implementation** for a future
PR.

* Moves the logic to keep the local `Configuration` updated from the
`AppConfigurationHttpClient` client into a dedicated
`network::live_configuration` module:
* Core logic to retrieve the configuration goes to `UpdateThreadWorker`
* Nice wrapper to handle offline modes goes to `LiveConfigurationImpl`
(**note that offline mode feature is not yet in this PR)
* `AppConfigurationIBMClient` is just a wrapper over
`AppConfigurationHttpClient` using a _default` live configuration.
* New `utils/thread_handle.rs` file with a nice utility to implement
scoped-threads and be able to check its status
 * New functionality has high test coverage 🧪 

---

Technical debt (added to #60 description):
- Review `pub` visibility of `models.rs` (basically go through all the
crate again) --> If we don't expose the `LiveConfiguration` it will be
possible to make all of them _private_.
- Review traits `PartialEq` and `Eq` added to `models.rs`
- Are the tests in `app_configuration_client_http.rs` actually testing
something?
- Go through all the errors hierarchy

---------

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>