feat: Introduce `ConfigurationProvider` trait (#74)
I want to hide the concept of a `Configuration` as an implementation
detail and instead introduce a trait (`ConfigurationProvider`) that
offers the functionality to provide features and properties. This trait
will then be implemented by many of the structs we already have:
app-configuration clients, the `Configuration` itself,... and it will be
easier to mock (/dependency injection).

I also think this will help us to narrow the API of the crate: we can
hide `Configuration` and some other structs together with it.

---------

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>