chore: Renames associated to `SegmentRule` and `TargetingRule` (#80)
* For the JSON fields (`appconfiguration::models`): the name in Rust
matches the name in the JSON. There is no need to rename, the more they
look like each other, the easier it will be to understand. We can
(should) document them a little bit, but they will be _private_ (🤞 ), so
it should be internal documentation only.
* In the `segment_evaluation` module, I changed the names to
`TargetingRule`, as suggested by @rainerschoe is previous comments.
Better documentation is still pending.

---------

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>