refact: Hide `AppConfigurationClientHttp` (and others) from the user  (#81)
I realized that the main blocker we had to hide most of our API was the
`AppConfigurationClientHttp`. This API of this object was allowing the
user to provide something that exposed the `Configuration` type. That's
all. With this in mind, there is no way we can hide the `Configuration`.

However, if we have a look, the only purpose of
`AppConfigurationClietHttp` is to provide a way to inject behaviour, and
right now we are only using it to write our integration tests. **We
cannot make something public just because we need it to write tests**.

So, here are the main changes:
 * Hide `AppConfigurationClientHttp`
* Provide a new `test_utils` module that basically calls the
`AppConfigurationClientHttp` ctor and returns the interface.
 * Now we can make `Configuration` private 😃

---------

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>