refact: `network` module returns `Configuration` object (#85)
Two main changes in this PR:
* `network` module returns a `Configuration` object. When we run
`ServerClient::get_configuration()` we already know which one is the
`environment_id`, so we can use it to retrieve the `Configuration`
object instead of the `ConfigurationJSON` one. **We hide JSON models
from external modules**, we use our models in the (internal) interfaces.
* Indeed, now the only consumer of `ConfigurationJSON` is our internal
model `Configuration`. _Note.- It should probably be the other way
around, JSON knows about the internal models, but that's a different
change_.

---------

Signed-off-by: Javier G. Sogo <jgsogo@gmail.com>